Separate out some common functions from the semgrex interface
