+++ stanza/server/semgrex.py
@@ -32,24 +32,9 @@ for multiple queries, but we didn't do any of those.  We do, however,
 accept pull requests...
 """
 
-import subprocess
-
 import stanza
 from stanza.protobuf import SemgrexRequest, SemgrexResponse
-from stanza.server.client import resolve_classpath
-
-def send_request(request, response_type, java_main):
-    """
-    Use subprocess to run the Semgrex processor on the given request
-
-    Returns the protobuf response
-    """
-    pipe = subprocess.run(["java", "-cp", resolve_classpath(), java_main],
-                          input=request.SerializeToString(),
-                          stdout=subprocess.PIPE)
-    response = response_type()
-    response.ParseFromString(pipe.stdout)
-    return response
+from stanza.server.java_protobuf_requests import send_request, add_token, add_word_to_graph
 
 def send_semgrex_request(request):
     return send_request(request, SemgrexResponse,
@@ -70,27 +55,8 @@ def process_doc(doc, *semgrex_patterns):
         word_idx = 0
         for token in sentence.tokens:
             for word in token.words:
-                query_token = query.token.add()
-                query_token.word = word.text
-                query_token.value = word.text
-                if word.lemma is not None:
-                    query_token.lemma = word.lemma
-                if word.xpos is not None:
-                    query_token.pos = word.xpos
-                if word.upos is not None:
-                    query_token.coarseTag = word.upos
-                if token.ner is not None:
-                    query_token.ner = token.ner
-
-                node = query.graph.node.add()
-                node.sentenceIndex = sent_idx+1
-                node.index = word_idx+1
-
-                if word.head != 0:
-                    edge = query.graph.edge.add()
-                    edge.source = word.head
-                    edge.target = word_idx+1
-                    edge.dep = word.deprel
+                add_token(query.token, word, token)
+                add_word_to_graph(query.graph, word, sent_idx, word_idx)
 
                 word_idx = word_idx + 1
